Lots of reorganizing, starting clustering.
- Moved things around to allow for packaging.
- Creating clustering file.
- Things maybe still work.
