+++ requirements.txt
@@ -2,7 +2,6 @@ boto==2.43.0
 bz2file==0.98
 funcsigs==1.0.2
 gensim==0.13.3
-Keras==1.1.1
 mock==2.0.0
 numpy==1.11.2
 pandas==0.19.1
@@ -19,6 +18,4 @@ six==1.10.0
 sklearn==0.0
 smart-open==1.3.5
 stop-words==2015.2.23.1
-tensorflow==0.11.0rc2
-Theano==0.8.2
 wordcloud==1.2.1
