+++ setup.py
@@ -1,0 +1,13 @@
+import os
+from setuptools import setup
+
+setup(
+    name="modelling",
+    version='0.0.1',
+    url="https://github.com/migrantnewsfilter/Modelling",
+
+    py_modules=['modelling'],
+    zip_safe=False,
+
+    install_requires=[]
+)
